refactor tests to demonstrate intent
