+++ src/app/views/sidebar/resource-explorer/panels/postman.util.ts
@@ -1,17 +1,10 @@
 import { Guid } from 'guid-typescript';
+
 import { IPostmanCollection, Item } from '../../../../../types/postman-collection';
-import { IResourceLink, IResourceLabel, IResourceMethod } from '../../../../../types/resources';
+import { IResourceLink } from '../../../../../types/resources';
 import { GRAPH_URL } from '../../../../services/graph-constants';
-import { downloadToLocal } from '../../../../utils/download';
-import { flatten, getUrlFromLink } from '../resource-explorer.utils';
 
-export function exportCollection(paths: IResourceLink[]) {
-  const content = generatePostmanCollection(paths);
-  const filename = `${content.info.name}-${content.info._postman_id}.postman_collection.json`;
-  downloadToLocal(content, filename);
-}
-
-function generatePostmanCollection(paths: IResourceLink[]): IPostmanCollection {
+export function generatePostmanCollection(paths: IResourceLink[]): IPostmanCollection {
   const collection: IPostmanCollection = {
     info: {
       _postman_id: Guid.create().toString(),
@@ -53,26 +46,4 @@ function generateItemsFromPaths(resources: IResourceLink[]): Item[] {
     list.push(item);
   });
   return list;
-}
-
-export function getResourcePaths(item: IResourceLink, version: string): IResourceLink[] {
-  const { links } = item;
-  const content: IResourceLink[] = flatten(links).filter((k: any) => k.type === 'path');
-  content.unshift(item);
-  if (content.length > 0) {
-    content.forEach((element: IResourceLink) => {
-      const methods = element.labels.find((k: IResourceLabel) => k.name === version)?.methods || [];
-      const listOfMethods: IResourceMethod[] = [];
-      methods.forEach((method: string) => {
-        listOfMethods.push({
-          name: method.toUpperCase(),
-          checked: true
-        });
-      });
-      element.version = version;
-      element.url = `${getUrlFromLink(element)}`;
-      element.methods = listOfMethods;
-    });
-  }
-  return content;
 }
